+++ packages/rax-miniapp-runtime-webpack-plugin/src/platforms/wechat.js
@@ -26,6 +26,7 @@ const View = {
     'hover-start-time': '50',
     'hover-stay-time': '400',
     'hover-stop-propagation': 'false',
+    animation: ''
   },
   events: {
     ...animationEvents
@@ -38,8 +39,38 @@ const View = {
 
 const CatchView = Object.assign({}, View);
 
+const StaticView = {
+  props: {
+    'hover-class': addSingleQuote('none'),
+    'hover-start-time': '50',
+    'hover-stay-time': '400',
+    'hover-stop-propagation': 'false',
+    animation: ''
+  }
+};
+
+const PureView = {};
+
+const NoTouchView = {
+  props: {
+    'hover-class': addSingleQuote('none'),
+    'hover-start-time': '50',
+    'hover-stay-time': '400',
+    'hover-stop-propagation': 'false',
+    animation: ''
+  },
+  events: {
+    ...animationEvents
+  },
+  basicEvents: {
+    ...tapEvents
+  }
+};
+
 const HElement = {
-  props: {},
+  props: {
+    animation: ''
+  },
   basicEvents: {
     ...tapEvents,
     ...touchEvents
@@ -47,6 +78,17 @@ const HElement = {
 };
 
 const CatchHElement = Object.assign({}, HElement);
+
+const PureHElement = {};
+
+const NoTouchHElement = {
+  props: {
+    animation: ''
+  },
+  basicEvents: {
+    ...tapEvents
+  }
+};
 
 const HComment = {};
 
@@ -186,8 +228,16 @@ const Text = {
     'number-of-lines': ''
   },
   basicEvents: {
-    ...tapEvents,
-    ...touchEvents
+    ...tapEvents
+  }
+};
+
+const StaticText = {
+  props: {
+    selectable: 'false',
+    space: '',
+    decode: 'false',
+    'number-of-lines': ''
   }
 };
 
@@ -249,7 +299,8 @@ const Button = {
     'send-message-path': '',
     'send-message-img': '',
     'app-parameter': '',
-    'show-message-card': 'false'
+    'show-message-card': 'false',
+    'data-params': '' // For share button
   },
   events: {
     GetUserInfo: '',
@@ -458,9 +509,18 @@ const Picker = {
     'header-text': '',
     disabled: 'false',
     mode: addSingleQuote('selector'),
+    range: '[]',
+    'range-key': '',
+    value: '',
+    start: '',
+    end: '',
+    fields: '',
+    'custom-item': ''
   },
   events: {
+    Change: '',
     Cancel: '',
+    ColumnChange: ''
   },
 };
 
@@ -506,6 +566,16 @@ const Image = {
   basicEvents: {
     ...tapEvents,
     ...touchEvents
+  }
+};
+
+const StaticImage = {
+  props: {
+    src: '',
+    mode: addSingleQuote('scaleToFill'),
+    webp: 'false',
+    'lazy-load': 'false',
+    'show-menu-by-longpress': 'false',
   }
 };
 
@@ -752,6 +822,9 @@ const OfficialAccount = {
 exports.internalComponents = {
   View,
   CatchView,
+  StaticView,
+  PureView,
+  NoTouchView,
   Swiper,
   SwiperItem,
   ScrollView,
@@ -760,6 +833,7 @@ exports.internalComponents = {
   MovableView,
   MovableArea,
   Text,
+  StaticText,
   Icon,
   Progress,
   RichText,
@@ -780,6 +854,7 @@ exports.internalComponents = {
   Picker,
   Navigator,
   Image,
+  StaticImage,
   Video,
   Canvas,
   Camera,
@@ -791,86 +866,95 @@ exports.internalComponents = {
   OfficialAccount,
   HElement,
   CatchHElement,
+  PureHElement,
+  NoTouchHElement,
   HComment
 };
 
 exports.derivedComponents = new Map([
-  ['catch-view', 'view'],
-  ['catch-h-element', 'view'],
-  ['h-element', 'view'],
-  ['h-comment', 'block']
+  ['CatchView', 'View'],
+  ['StaticView', 'View'],
+  ['PureView', 'View'],
+  ['NoTouchView', 'View'],
+  ['CatchHElement', 'View'],
+  ['PureHElement', 'View'],
+  ['NoTouchHElement', 'View'],
+  ['HElement', 'View'],
+  ['StaticText', 'Text'],
+  ['StaticImage', 'Image'],
+  ['HComment', 'Block']
 ]);
 
 exports.controlledComponents = new Set([
-  'input',
-  'checkbox',
-  'picker',
-  'picker-view',
-  'radio',
-  'slider',
-  'switch',
-  'textarea'
+  'Input',
+  'Checkbox',
+  'Picker',
+  'PickerView',
+  'Radio',
+  'Slider',
+  'Switch',
+  'Textarea'
 ]);
 
 exports.focusComponents = new Set([
-  'input',
-  'textarea'
+  'Input',
+  'Textarea'
 ]);
 
 exports.voidElements = new Set([
-  'h-comment'
+  'HComment'
 ]);
 
 exports.voidChildrenElements = new Set([
-  'progress',
-  'icon',
-  'rich-text',
-  'input',
-  'textarea',
-  'checkbox',
-  'radio',
-  'editor',
-  'slider',
-  'switch',
-  'live-pusher',
-  'h-comment',
-  'open-data',
-  'image',
-  'video',
-  'canvas',
-  'web-view',
-  'live-player',
-  'live-pusher',
-  'official-account'
+  'Progress',
+  'Icon',
+  'RichText',
+  'Input',
+  'Textarea',
+  'Checkbox',
+  'Radio',
+  'Editor',
+  'Slider',
+  'Switch',
+  'LivePusher',
+  'HComment',
+  'OpenData',
+  'Image',
+  'Video',
+  'Canvas',
+  'WebView',
+  'LivePlayer',
+  'LivePusher',
+  'OfficialAccount'
 ]);
 
 exports.nestElements = new Map([
-  ['view', -1],
-  ['cover-view', -1],
-  ['text', 6],
-  ['label', 6],
-  ['form', 4],
-  ['scroll-view', 4],
-  ['swiper', 4],
-  ['swiper-item', 4]
+  ['View', -1],
+  ['CoverView', -1],
+  ['Text', 6],
+  ['Label', 6],
+  ['Form', 4],
+  ['ScrollView', 4],
+  ['Swiper', 4],
+  ['SwiperItem', 4]
 ]);
 
 exports.shouldNotGenerateTemplateComponents = new Set([
-  'swiper-item',
-  'picker-view-column',
-  'movable-view'
+  'SwiperItem',
+  'PickerViewColumn',
+  'MovableView'
 ]);
 
 exports.needModifyChildrenComponents = {
-  swiper: (children, level) => `
+  Swiper: (children, level) => `
     <swiper-item wx:for="{{r.children}}" wx:if="{{item.nodeType !== 'h-comment'}}" wx:key="nodeId">
       <template is="{{tool.b(cid + 1)}}" data="{{r: item.children, c: tool.d(c, 'swiper')}}" />
     </swiper-item>`,
-  'movable-area': children => `
+  MovableArea: children => `
     <movable-view wx:for="{{r.children}}" wx:key="nodeId" wx:if="{{item.nodeType !== 'h-comment'}}" direction="{{item['direction']||'none'}}" inertia="{{tool.a(item['inertia'],false)}}" out-of-bounds="{{tool.a(item['out-of-bounds'],false)}}" x="{{tool.a(item['x'],0)}}" y="{{tool.a(item['y'],0)}}" damping="{{tool.a(item['damping'],20)}}" friction="{{tool.a(item['friction'],2)}}" disabled="{{tool.a(item['disabled'],false)}}" scale="{{tool.a(item['scale'],false)}}" scale-min="{{tool.a(item['scale-min'],0.5)}}" scale-max="{{tool.a(item['scale-max'],10)}}" scale-value="{{tool.a(item['scale-value'],1)}}" animation="{{tool.a(item['animation'],true)}}" bindchange="onMovableViewChange" bindscale="onMovableViewScale" bindtouchstart="onTouchStart" bindtouchmove="onTouchMove" bindtouchend="onTouchEnd" bindtouchcancel="onTouchCancel" bindlongtap="onLongTap" bindtap="onTap" style="{{item.style}}" class="{{item.class}}" id="{{item.id}}" data-private-node-id="{{item.nodeId}}">
       <template is="{{tool.b(cid + 1)}}" data="{{r: item.children, c: tool.d(c, 'movable-area')}}" />
     </movable-view>`,
-  'scroll-view': children => `
+  ScrollView: children => `
     <block wx:for="{{r.children}}" wx:key="nodeId">
       <block wx:if="{{item.nodeId}}">
         <template is="{{tool.c(item.nodeType, tool.d(c, 'scroll-view'))}}" data="{{r: item, c: tool.d(c, 'scroll-view'), cid: cid}}" />
@@ -880,10 +964,21 @@ exports.needModifyChildrenComponents = {
       </block>
     </block>
   `,
-  'picker-view': children => `
+  PickerView: children => `
     <picker-view-column wx:for="{{r.children}}" wx:key="nodeId" wx:if="{{item.nodeType !== 'h-comment'}}">
       <template is="{{tool.b(cid + 1)}}" data="{{r: item.children, c: tool.d(c, 'picker-view')}}" />
-    </picker-view-column>`
+    </picker-view-column>
+  `,
+  Map: children => `
+    <block wx:for="{{r.children}}" wx:key="nodeId">
+      <block wx:if="{{item.nodeId}}">
+        <template is="{{tool.c(item.nodeType, tool.d(c, 'map'))}}" data="{{r: item, c: tool.d(c, 'map'), cid: cid}}" />
+      </block>
+      <block wx:else>
+        <block>{{item.content}}</block>
+      </block>
+    </block>
+  `
 };
 
 exports.adapter = {