+++ packages/miniapp-builder-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.2] - 2021-03-30
+
+### Added
+
+- `autoInstallNpm` method for installing npm for native miniapp automatically
+
 ## [0.2.1] - 2021-03-11
 
 ### Added