+++ packages/miniapp-render/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Changelog
 
+## [2.3.1] - 2021-04-12
+
+### Fixed
+
+- `_processNodeType` error in normal components
+
+### Changed
+
+- Remove text from touch components
+
+## [2.3.0] - 2021-04-08
+
+### Added
+
+- Grade view/text/image/h-element into static/pure/no-touch/no-appear/no-appear-touch level
+
+### Changed
+
+- Add dispatch document modify callback
+- Remove some unnecessary element destory logic in page unload lifecycle
+- Remove some useless code
+
+
+## [2.2.3] - 2021-03-31
+
+### Changed
+
+- Share window when share memory in subpackages
+
+## [2.2.2] - 2021-03-30
+
+### Fixed
+
+- Miss app instance in native app licecycle methods
+
+## [2.2.1] - 2021-03-25
+
+### Fixed
+- Can't identify contact-button as builtin components
+- Avoid `requireModule` effect building in ali miniapp
+- Update dataset didn't trigger update
+## [2.2.0] - 2021-03-23
+
+### Added
+
+- Support inject native app licecycle methods
+- Add contact-button
+- Add `requireModule` method in app to support share memory in subpackages
+
+### Fixed
+- Miss official-account component import
+
 ## [2.1.1] - 2021-03-15
 
 ### Fixed